Explicitly get location before getting any weather data
Due to the way FF handles multiple requests before a user has shared location, only the last call to the geo api was getting resolved.
Even where this isn't an issue in Chrome, the geolocaiton position wasn't cached in the geoservice and a duplicate request was still made.
To eliminate this behavior, an explcit call is made to get the geolocation synchronously so any further weather calls have to wait.